+++ Components/WebSocket/Events/WebSocketDisconnected.cs
@@ -1,0 +1,11 @@
+﻿using Slipstream.Shared;
+
+namespace Slipstream.Components.WebSocket.Events
+{
+    public class WebSocketDisconnected : IEvent
+    {
+        public string EventType => nameof(WebSocketDisconnected);
+
+        public IEventEnvelope Envelope { get; set; } = new EventEnvelope();
+    }
+}