+++ Components/WebSocket/Lua/IWebSocketLuaReference.cs
@@ -1,0 +1,8 @@
+﻿using Slipstream.Shared.Lua;
+
+namespace Slipstream.Components.WebSocket.Lua
+{
+    public interface IWebSocketLuaReference : ILuaReference
+    {
+    }
+}